requirements.txt
openai-whisper
transformers
torch
spacy
nltk
flask
flask-sqlalchemy
deep-translator
gTTS
gunicorn
